Add decision criteria data to experiment api
