+++ packages/back-end/src/api/openapi/schemas/Experiment.yaml
@@ -226,3 +226,17 @@ properties:
     type: boolean
   hasURLRedirects:
     type: boolean
+  decisionCriteriaStatus:
+    type: object
+    required:
+      - status
+    properties:
+      status:
+        type: string
+        enum:
+          - "Running"
+          - "Stopped"
+          - "Draft"
+          - "Archived"
+      detailedStatus:
+        type: string
