+++ packages/back-end/src/validators/openapi.ts
@@ -40,7 +40,7 @@ export const apiFeatureExperimentRefRuleValidator = z.object({ "description": z.
 
 export const apiSdkConnectionValidator = z.object({ "id": z.string(), "dateCreated": z.string(), "dateUpdated": z.string(), "name": z.string(), "organization": z.string(), "languages": z.array(z.string()), "sdkVersion": z.string().optional(), "environment": z.string(), "project": z.string().describe("Use 'projects' instead. This is only for backwards compatibility and contains the first project only."), "projects": z.array(z.string()).optional(), "encryptPayload": z.boolean(), "encryptionKey": z.string(), "includeVisualExperiments": z.boolean().optional(), "includeDraftExperiments": z.boolean().optional(), "includeExperimentNames": z.boolean().optional(), "includeRedirectExperiments": z.boolean().optional(), "includeRuleIds": z.boolean().optional(), "key": z.string(), "proxyEnabled": z.boolean(), "proxyHost": z.string(), "proxySigningKey": z.string(), "sseEnabled": z.boolean().optional(), "hashSecureAttributes": z.boolean().optional(), "remoteEvalEnabled": z.boolean().optional(), "savedGroupReferencesEnabled": z.boolean().optional() }).strict()
 
-export const apiExperimentValidator = z.object({ "id": z.string(), "trackingKey": z.string(), "dateCreated": z.string(), "dateUpdated": z.string(), "name": z.string(), "type": z.enum(["standard","multi-armed-bandit"]), "project": z.string(), "hypothesis": z.string(), "description": z.string(), "tags": z.array(z.string()), "owner": z.string(), "archived": z.boolean(), "status": z.string(), "autoRefresh": z.boolean(), "hashAttribute": z.string(), "fallbackAttribute": z.string().optional(), "hashVersion": z.union([z.literal(1), z.literal(2)]), "disableStickyBucketing": z.boolean().optional(), "bucketVersion": z.coerce.number().optional(), "minBucketVersion": z.coerce.number().optional(), "variations": z.array(z.object({ "variationId": z.string(), "key": z.string(), "name": z.string(), "description": z.string(), "screenshots": z.array(z.string()) })), "phases": z.array(z.object({ "name": z.string(), "dateStarted": z.string(), "dateEnded": z.string(), "reasonForStopping": z.string(), "seed": z.string(), "coverage": z.coerce.number(), "trafficSplit": z.array(z.object({ "variationId": z.string(), "weight": z.coerce.number() })), "namespace": z.object({ "namespaceId": z.string(), "range": z.array(z.any()) }).optional(), "targetingCondition": z.string(), "prerequisites": z.array(z.object({ "id": z.string(), "condition": z.string() })).optional(), "savedGroupTargeting": z.array(z.object({ "matchType": z.enum(["all","any","none"]), "savedGroups": z.array(z.string()) })).optional() })), "settings": z.object({ "datasourceId": z.string(), "assignmentQueryId": z.string(), "experimentId": z.string(), "segmentId": z.string(), "queryFilter": z.string(), "inProgressConversions": z.enum(["include","exclude"]), "attributionModel": z.enum(["firstExposure","experimentDuration"]).describe("Setting attribution model to `\"experimentDuration\"` is the same as selecting \"Ignore Conversion Windows\" for the Conversion Window Override."), "statsEngine": z.enum(["bayesian","frequentist"]), "regressionAdjustmentEnabled": z.boolean().optional(), "sequentialTestingEnabled": z.boolean().optional(), "sequentialTestingTuningParameter": z.coerce.number().optional(), "goals": z.array(z.object({ "metricId": z.string(), "overrides": z.object({ "delayHours": z.coerce.number().optional(), "windowHours": z.coerce.number().optional(), "window": z.enum(["conversion","lookback",""]).optional(), "winRiskThreshold": z.coerce.number().optional(), "loseRiskThreshold": z.coerce.number().optional() }) })), "secondaryMetrics": z.array(z.object({ "metricId": z.string(), "overrides": z.object({ "delayHours": z.coerce.number().optional(), "windowHours": z.coerce.number().optional(), "window": z.enum(["conversion","lookback",""]).optional(), "winRiskThreshold": z.coerce.number().optional(), "loseRiskThreshold": z.coerce.number().optional() }) })), "guardrails": z.array(z.object({ "metricId": z.string(), "overrides": z.object({ "delayHours": z.coerce.number().optional(), "windowHours": z.coerce.number().optional(), "window": z.enum(["conversion","lookback",""]).optional(), "winRiskThreshold": z.coerce.number().optional(), "loseRiskThreshold": z.coerce.number().optional() }) })), "activationMetric": z.object({ "metricId": z.string(), "overrides": z.object({ "delayHours": z.coerce.number().optional(), "windowHours": z.coerce.number().optional(), "window": z.enum(["conversion","lookback",""]).optional(), "winRiskThreshold": z.coerce.number().optional(), "loseRiskThreshold": z.coerce.number().optional() }) }).optional() }), "resultSummary": z.object({ "status": z.string(), "winner": z.string(), "conclusions": z.string(), "releasedVariationId": z.string(), "excludeFromPayload": z.boolean() }).optional(), "shareLevel": z.enum(["public","organization"]).optional(), "publicUrl": z.string().optional(), "banditScheduleValue": z.coerce.number().optional(), "banditScheduleUnit": z.enum(["days","hours"]).optional(), "banditBurnInValue": z.coerce.number().optional(), "banditBurnInUnit": z.enum(["days","hours"]).optional(), "linkedFeatures": z.array(z.string()).optional(), "hasVisualChangesets": z.boolean().optional(), "hasURLRedirects": z.boolean().optional() }).strict()
+export const apiExperimentValidator = z.object({ "id": z.string(), "trackingKey": z.string(), "dateCreated": z.string(), "dateUpdated": z.string(), "name": z.string(), "type": z.enum(["standard","multi-armed-bandit"]), "project": z.string(), "hypothesis": z.string(), "description": z.string(), "tags": z.array(z.string()), "owner": z.string(), "archived": z.boolean(), "status": z.string(), "autoRefresh": z.boolean(), "hashAttribute": z.string(), "fallbackAttribute": z.string().optional(), "hashVersion": z.union([z.literal(1), z.literal(2)]), "disableStickyBucketing": z.boolean().optional(), "bucketVersion": z.coerce.number().optional(), "minBucketVersion": z.coerce.number().optional(), "variations": z.array(z.object({ "variationId": z.string(), "key": z.string(), "name": z.string(), "description": z.string(), "screenshots": z.array(z.string()) })), "phases": z.array(z.object({ "name": z.string(), "dateStarted": z.string(), "dateEnded": z.string(), "reasonForStopping": z.string(), "seed": z.string(), "coverage": z.coerce.number(), "trafficSplit": z.array(z.object({ "variationId": z.string(), "weight": z.coerce.number() })), "namespace": z.object({ "namespaceId": z.string(), "range": z.array(z.any()) }).optional(), "targetingCondition": z.string(), "prerequisites": z.array(z.object({ "id": z.string(), "condition": z.string() })).optional(), "savedGroupTargeting": z.array(z.object({ "matchType": z.enum(["all","any","none"]), "savedGroups": z.array(z.string()) })).optional() })), "settings": z.object({ "datasourceId": z.string(), "assignmentQueryId": z.string(), "experimentId": z.string(), "segmentId": z.string(), "queryFilter": z.string(), "inProgressConversions": z.enum(["include","exclude"]), "attributionModel": z.enum(["firstExposure","experimentDuration"]).describe("Setting attribution model to `\"experimentDuration\"` is the same as selecting \"Ignore Conversion Windows\" for the Conversion Window Override."), "statsEngine": z.enum(["bayesian","frequentist"]), "regressionAdjustmentEnabled": z.boolean().optional(), "sequentialTestingEnabled": z.boolean().optional(), "sequentialTestingTuningParameter": z.coerce.number().optional(), "goals": z.array(z.object({ "metricId": z.string(), "overrides": z.object({ "delayHours": z.coerce.number().optional(), "windowHours": z.coerce.number().optional(), "window": z.enum(["conversion","lookback",""]).optional(), "winRiskThreshold": z.coerce.number().optional(), "loseRiskThreshold": z.coerce.number().optional() }) })), "secondaryMetrics": z.array(z.object({ "metricId": z.string(), "overrides": z.object({ "delayHours": z.coerce.number().optional(), "windowHours": z.coerce.number().optional(), "window": z.enum(["conversion","lookback",""]).optional(), "winRiskThreshold": z.coerce.number().optional(), "loseRiskThreshold": z.coerce.number().optional() }) })), "guardrails": z.array(z.object({ "metricId": z.string(), "overrides": z.object({ "delayHours": z.coerce.number().optional(), "windowHours": z.coerce.number().optional(), "window": z.enum(["conversion","lookback",""]).optional(), "winRiskThreshold": z.coerce.number().optional(), "loseRiskThreshold": z.coerce.number().optional() }) })), "activationMetric": z.object({ "metricId": z.string(), "overrides": z.object({ "delayHours": z.coerce.number().optional(), "windowHours": z.coerce.number().optional(), "window": z.enum(["conversion","lookback",""]).optional(), "winRiskThreshold": z.coerce.number().optional(), "loseRiskThreshold": z.coerce.number().optional() }) }).optional() }), "resultSummary": z.object({ "status": z.string(), "winner": z.string(), "conclusions": z.string(), "releasedVariationId": z.string(), "excludeFromPayload": z.boolean() }).optional(), "shareLevel": z.enum(["public","organization"]).optional(), "publicUrl": z.string().optional(), "banditScheduleValue": z.coerce.number().optional(), "banditScheduleUnit": z.enum(["days","hours"]).optional(), "banditBurnInValue": z.coerce.number().optional(), "banditBurnInUnit": z.enum(["days","hours"]).optional(), "linkedFeatures": z.array(z.string()).optional(), "hasVisualChangesets": z.boolean().optional(), "hasURLRedirects": z.boolean().optional(), "decisionCriteriaStatus": z.object({ "status": z.enum(["Running","Stopped","Draft","Archived"]), "detailedStatus": z.string().optional() }).optional() }).strict()
 
 export const apiExperimentSnapshotValidator = z.object({ "id": z.string(), "experiment": z.string(), "status": z.string() }).strict()
