+++ packages/front-end/components/Experiment/TabbedPage/ExperimentStatusIndicator.tsx
@@ -1,10 +1,8 @@
 import { Flex, Tooltip } from "@radix-ui/themes";
 import { ExperimentDataForStatusStringDates } from "back-end/types/experiment";
+import { StatusIndicatorData } from "shared/enterprise";
 import Badge from "@/components/Radix/Badge";
-import {
-  StatusIndicatorData,
-  useExperimentStatusIndicator,
-} from "@/hooks/useExperimentStatusIndicator";
+import { useExperimentStatusIndicator } from "@/hooks/useExperimentStatusIndicator";
 
 type LabelFormat = "full" | "status-only" | "detail-only";
