+++ packages/shared/src/enterprise/decision-criteria/decisionCriteria.ts
@@ -23,6 +23,10 @@ import {
 } from "../../constants";
 import { daysBetween } from "../../dates";
 import { getMultipleExposureHealthData, getSRMHealthData } from "../../health";
+import {
+  DEFAULT_DECISION_CRITERIA,
+  DEFAULT_DECISION_CRITERIAS,
+} from "./constants";
 
 // Evaluate a single rule on a variation result
 // Returns the action if the rule is met, otherwise undefined
@@ -460,4 +464,14 @@ export function getExperimentResultStatus({
       return daysLeftStatus;
     }
   }
+}
+
+export function getDefaultDecisionCriteriaForOrg(
+  settings?: OrganizationSettings
+) {
+  return !settings?.defaultDecisionCriteriaId
+    ? DEFAULT_DECISION_CRITERIA
+    : DEFAULT_DECISION_CRITERIAS.find(
+        (dc) => dc.id === settings.defaultDecisionCriteriaId
+      );
 }
