+++ packages/shared/src/enterprise/decision-criteria/index.ts
@@ -1,2 +1,3 @@
 export * from "./decisionCriteria";
 export * from "./constants";
+export * from "./statusIndicatorData";
